tools/requirements.txt
# Core Dependencies
python-dotenv==1.0.0
requests==2.31.0

# Supabase
supabase==2.3.4

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.41.2

# RSS Feed Parsing
feedparser==6.0.11

# Reddit API
praw==7.7.1

# Twitter/Social Media
tweepy==4.14.0

# Data Processing
python-dateutil==2.8.2

# Testing (optional)
pytest==7.4.0
pytest-asyncio==0.23.4
